Fixed Rotation, Animated Camera, Added Space Station ++
Rotation now works properly and the issue with the dial has been fixed. Camera animates into position when rotating rather than snapping, but it is a bit buggy where the rotation sometimes won't be smooth and sometimes it won't choose the shortest path. Space Stations are now in game and fully functional. Added some temporary changes to UI and tweaked the health script to include the shield.